[Rupesh-patil].Updated NoteServiceImpl

+++ src/main/java/com/bridgelabz/notes/service/NoteServiceImpl.java
@@ -6,12 +6,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
 import com.bridgelabz.notes.dao.INotesDao;
 import com.bridgelabz.notes.model.Note;
 import com.bridgelabz.notes.response.NoteCreatedResponse;
 import com.bridgelabz.notes.response.NotesResponce;
 import com.bridgelabz.notes.util.JwtTokenUtil;
+import com.bridgelabz.notes.util.UserData;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -22,24 +24,27 @@ public class NoteServiceImpl {
 	JwtTokenUtil tokenUtil;
 	@Autowired
 	INotesDao noteDao;
+	@Autowired
+	RestTemplate restTemplate;
 
 	@Autowired
 	NoteCreatedResponse notCreatedResponse;
 	@Autowired
 	NotesResponce noteObjectResponse;
-	String status400="No notes Available or UserToken Expired";
+	String status400 = "No notes Available";
+	String userNotExist = "User Token Expired or User Does Note Exist";
+
 	@Transactional
 	public ResponseEntity<NotesResponce> createNote(Note note, String token) {
-		Long id = getUserIdFromToken(token);
-		Note noteTemp = noteDao.createNote(id, note);
-		if (id != 0) {
-			if (noteTemp != null)
+		if (verifyUser(token)) {
+			if (noteDao.createNote(getUserIdFromToken(token), note) != null)
 				return ResponseEntity.status(HttpStatus.CREATED).body(new NotesResponce(201, "Note Created"));
 			else
 				return ResponseEntity.status(HttpStatus.FAILED_DEPENDENCY)
 						.body(new NotesResponce(424, "Some internal Errors !!Be patients we will Solve Sorty"));
+		}else {
+			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new NotesResponce(400, "User Does Not exist"));
 		}
-		return ResponseEntity.status(HttpStatus.BAD_GATEWAY).body(new NotesResponce(424, "USer Does not exist"));
 	}
 
 	public long getUserIdFromToken(String token) {
@@ -51,16 +56,14 @@ public class NoteServiceImpl {
 		log.info("abccc");
 		try {
 			Note note = noteDao.getNoteById(noteId);
-			Long userId = getUserIdFromToken(token);
-			if (userId != 0) {
-				if (noteDTO.getTitle() != null)
-					note.setTitle(noteDTO.getTitle());
-				if (noteDTO.getDescription() != null)
-					note.setDescription(noteDTO.getDescription());
-				if (noteDao.updateNote(noteId, note, userId) != null)
-					return ResponseEntity.status(HttpStatus.OK).body(new NotesResponce(200, "Note Updated"));
+			if (verifyUser(token)) {
+				note.setTitle(noteDTO.getTitle());
+				note.setDescription(noteDTO.getDescription());
+				noteDao.updateNote(noteId, note, getUserIdFromToken(token));
+				return ResponseEntity.status(HttpStatus.OK).body(new NotesResponce(200, "Note Updated"));
+			} else {
+				return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new NotesResponce(204, userNotExist));
 			}
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new NotesResponce(400, "Note Can't Updated"));
 		} catch (Exception e) {
 			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new NotesResponce(400, "User Does Not exist"));
 		}
@@ -70,10 +73,15 @@ public class NoteServiceImpl {
 	public ResponseEntity<String> delete(Long noteId, String token) {
 		Long userId = getUserIdFromToken(token);
 		try {
-			if (noteDao.deleteNote(userId, noteId) == 1) {
-				return ResponseEntity.status(HttpStatus.ACCEPTED).body("Note deleted SuccessFully");
+			if (verifyUser(token)) {
+				if (noteDao.deleteNote(userId, noteId) == 1)
+					return ResponseEntity.status(HttpStatus.ACCEPTED).body("Note deleted SuccessFully");
+				else
+					return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Notes unavailable");
+
 			} else {
-				return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Notes unavailable");
+				return ResponseEntity.status(HttpStatus.NO_CONTENT).body("User Does NOt Exist");
+
 			}
 		} catch (NullPointerException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("User Does Note Exist" + e);
@@ -84,15 +92,19 @@ public class NoteServiceImpl {
 	public ResponseEntity<NotesResponce> getNoteList(String token) {
 		log.info("Number of Notes Available" + noteDao.getAllNoteList(getUserIdFromToken(token)).size());
 		try {
-			if (noteDao.getAllNoteList(getUserIdFromToken(token)).isEmpty())
-				return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-						.body(new NotesResponce(400, "No any Notes Available"));
-			else
-				return ResponseEntity.status(HttpStatus.FOUND)
-						.body(new NotesResponce(200,
-								"Number of Notes Available Are: "
-										+ noteDao.getAllNoteList(getUserIdFromToken(token)).size(),
-								noteDao.getAllNoteList(getUserIdFromToken(token))));
+			if (verifyUser(token)) {
+				if (noteDao.getAllNoteList(getUserIdFromToken(token)).isEmpty())
+					return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+							.body(new NotesResponce(400, "No any Notes Available"));
+				else
+					return ResponseEntity.status(HttpStatus.FOUND)
+							.body(new NotesResponce(200,
+									"Number of Notes Available Are: "
+											+ noteDao.getAllNoteList(getUserIdFromToken(token)).size(),
+									noteDao.getAllNoteList(getUserIdFromToken(token))));
+			} else {
+				return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new NotesResponce(204, userNotExist));
+			}
 		} catch (NullPointerException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
 					.body(new NotesResponce(400, "Pin Operation Sucessfully done"));
@@ -102,55 +114,64 @@ public class NoteServiceImpl {
 	@Transactional
 	public ResponseEntity<NotesResponce> pinnedNotes(String token, Long noteId) {
 		try {
-			Long userId = getUserIdFromToken(token);
-			Integer result=noteDao.pinUnpinNote(userId, noteId);
-			if (result == 1) {
-				return ResponseEntity.status(HttpStatus.ACCEPTED)
-						.body(new NotesResponce(202, "Note Pinned Sucessfully"));
-			}else if(result == 2){
-				return ResponseEntity.status(HttpStatus.ACCEPTED)
-						.body(new NotesResponce(202, "Note Unpinned Sucessfully"));
+			if (verifyUser(token)) {
+				Long userId = getUserIdFromToken(token);
+				Integer result = noteDao.pinUnpinNote(userId, noteId);
+				if (result == 1)
+					return ResponseEntity.status(HttpStatus.ACCEPTED)
+							.body(new NotesResponce(202, "Note Pinned Sucessfully"));
+				else if (result == 2)
+					return ResponseEntity.status(HttpStatus.ACCEPTED)
+							.body(new NotesResponce(202, "Note Unpinned Sucessfully"));
+				else
+					return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new NotesResponce(400, status400));
+
 			} else {
-				return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-						.body(new NotesResponce(400, status400));
+				return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new NotesResponce(204, userNotExist));
 			}
 		} catch (Exception e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-					.body(new NotesResponce(400, status400));
+					.body(new NotesResponce(400, status400 + "or User Token Expired Or UserDoes Not Exist"));
 		}
 	}
 
 	@Transactional
 	public ResponseEntity<NotesResponce> trashedNote(String token, Long noteId) {
 		try {
-			getUserIdFromToken(token);
-			Integer result=noteDao.trashedNote(noteId);
-			if (result == 1) {
-				return ResponseEntity.status(HttpStatus.ACCEPTED)
-						.body(new NotesResponce(202, "Note successfully moved to Trashed"));
-			} else if (result == 2) {
-				return ResponseEntity.status(HttpStatus.ACCEPTED)
-						.body(new NotesResponce(202, "Note successfully Restored"));
-			} else {
-				return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-						.body(new NotesResponce(400, status400));
-			}
+			if (verifyUser(token)) {
+				Integer result = noteDao.trashedNote(noteId);
+				if (result == 1)
+					return ResponseEntity.status(HttpStatus.ACCEPTED)
+							.body(new NotesResponce(202, "Note successfully moved to Trashed"));
+				else if (result == 2)
+					return ResponseEntity.status(HttpStatus.ACCEPTED)
+							.body(new NotesResponce(202, "Note successfully Restored"));
+				else
+					return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new NotesResponce(400, status400));
+
+			} else
+				return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new NotesResponce(204, userNotExist));
+
 		} catch (Exception e) {
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-					.body(new NotesResponce(400, status400));
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new NotesResponce(400, status400));
 		}
 	}
-	
+
 	public ResponseEntity<NotesResponce> getTrashNotes(String token) {
 		try {
-			return ResponseEntity.status(HttpStatus.ACCEPTED)
-					.body(new NotesResponce(202,
-							"Total Trashed Notes Are:" + noteDao.getTrashedNoteList(getUserIdFromToken(token)).size(),
-							noteDao.getTrashedNoteList(getUserIdFromToken(token))));
+			if (verifyUser(token))
+				return ResponseEntity.status(HttpStatus.ACCEPTED)
+						.body(new NotesResponce(202,
+								"Total Trashed Notes Are:"
+										+ noteDao.getTrashedNoteList(getUserIdFromToken(token)).size(),
+								noteDao.getTrashedNoteList(getUserIdFromToken(token))));
+			else
+
+				return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new NotesResponce(204, userNotExist));
+
 		} catch (Exception e) {
 			log.info("get Trashed" + e);
-			return ResponseEntity.status(HttpStatus.CONFLICT)
-					.body(new NotesResponce(400, status400 ));
+			return ResponseEntity.status(HttpStatus.CONFLICT).body(new NotesResponce(400, status400));
 		}
 	}
 
@@ -158,40 +179,49 @@ public class NoteServiceImpl {
 	public ResponseEntity<NotesResponce> archivedNote(String token, Long noteId) {
 		try {
 			getUserIdFromToken(token);
-			Integer result=noteDao.archivedNote(noteId);
-			if (result== 1) {
-				return ResponseEntity.status(HttpStatus.ACCEPTED)
-						.body(new NotesResponce(202, "Note Successfully moved in Achive Folder"));
-			} else if (result == 2) {
-				return ResponseEntity.status(HttpStatus.ACCEPTED)
-						.body(new NotesResponce(202, "Note Successfully moved from Achive Folder to DashBoard"));
+			Integer result = noteDao.archivedNote(noteId);
+			if (verifyUser(token)) {
+				if (result == 1)
+					return ResponseEntity.status(HttpStatus.ACCEPTED)
+							.body(new NotesResponce(202, "Note Successfully moved in Achive Folder"));
+				else if (result == 2)
+					return ResponseEntity.status(HttpStatus.ACCEPTED)
+							.body(new NotesResponce(202, "Note Successfully moved from Achive Folder to DashBoard"));
+				else
+					return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new NotesResponce(400, status400));
+
 			} else {
-				return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-						.body(new NotesResponce(400, status400));
+				return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new NotesResponce(204, userNotExist));
 			}
 		} catch (Exception e) {
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-					.body(new NotesResponce(400, status400));
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new NotesResponce(400, status400));
 		}
 	}
-	
-	
+
 	public ResponseEntity<NotesResponce> getArchivedNotes(String token) {
-		try {
-			return ResponseEntity.status(HttpStatus.ACCEPTED)
-					.body(new NotesResponce(202,
-							"Total Achived Notes Are:" + noteDao.getArchiveNoteList(getUserIdFromToken(token)).size(),
-							noteDao.getTrashedNoteList(getUserIdFromToken(token))));
-		} catch (Exception e) {
-			log.info("get Archived" + e);
-			return ResponseEntity.status(HttpStatus.CONFLICT)
-					.body(new NotesResponce(400, status400 ));
+		if (verifyUser(token)) {
+			try {
+				return ResponseEntity.status(HttpStatus.ACCEPTED)
+						.body(new NotesResponce(202,
+								"Total Achived Notes Are:"
+										+ noteDao.getArchiveNoteList(getUserIdFromToken(token)).size(),
+								noteDao.getTrashedNoteList(getUserIdFromToken(token))));
+			} catch (Exception e) {
+				log.info("get Archived" + e);
+				return ResponseEntity.status(HttpStatus.CONFLICT).body(new NotesResponce(400, status400));
+			}
+		} else {
+			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(new NotesResponce(204, userNotExist));
 		}
 	}
-	
-//	public  ResponseEntity<NotesResponce> getAllNotesByLabel(String token,String labelNAme)
-//	{
-//		
-//	}
+
+	public boolean verifyUser(String token) {
+		log.info("-------->>>>>>>>>>>>>Calling USerApi From NotesApi<<<<<<<<<<<<<<<<--------------------");
+		UserData userData = restTemplate.getForObject("http://localhost:8086/users/" + token, UserData.class);
+		log.info("--------->>>>>>>>>>>>Accessing DataFrom UserApi<<<<<<<<<<<---------------------");
+		log.info("verifyUserApi Using RestTemplate From UserApi Success--------->:"
+				+ (userData.getId() == getUserIdFromToken(token)));
+		return (userData.getId() == getUserIdFromToken(token));
+	}
 
 }
